requirements.txt
xlsxwriter
python-dotenv==0.17.0
requests==2.25.1